webdev_requirements.txt
# very basic list of web dev utilities for mid size projects
Django
django-tinymce
#django-endless-pagination
django-el-pagination
django-grappelli
django-filebrowser>=3.4.0
django-extensions
sphinx
Celery
South
#django-sentry
django-celery
django-kombu
supervisor
python-dateutil
mimeparse>=0.1.3
django-tastypie
